[chore] #3 global -> infra 폴더링 변경
