+++ src/main/java/com/ggang/be/api/user/controller/UserController.java
@@ -11,8 +11,8 @@ import com.ggang.be.api.user.NicknameValidator;
 import com.ggang.be.api.user.dto.*;
 import com.ggang.be.api.user.service.UserService;
 import com.ggang.be.domain.user.dto.UserSchoolDto;
-import com.ggang.be.global.infra.service.AppleLoginService;
-import com.ggang.be.global.infra.service.KakaoLoginService;
+import com.ggang.be.infra.service.AppleLoginService;
+import com.ggang.be.infra.service.KakaoLoginService;
 import com.ggang.be.global.jwt.JwtService;
 import com.ggang.be.global.jwt.TokenVo;
 import com.ggang.be.global.util.LengthValidator;
