+++ src/main/java/com/ggang/be/infra/Auth.java
@@ -1,4 +1,4 @@
-package com.ggang.be.global.infra;
+package com.ggang.be.infra;
 
 public record Auth(String platformId) {
 }
