+++ src/main/java/com/ggang/be/infra/oauth/AppleProperties.java
@@ -1,4 +1,4 @@
-package com.ggang.be.global.infra.oauth;
+package com.ggang.be.infra.oauth;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -8,8 +8,11 @@ import org.springframework.context.annotation.Configuration;
 @Getter
 @Setter
 @Configuration
-@ConfigurationProperties(prefix = "kakao")
-public class KakaoProperties {
+@ConfigurationProperties(prefix = "apple")
+public class AppleProperties {
+    private String teamId;
+    private String loginKey;
     private String clientId;
     private String redirectUri;
+    private String keyPath;
 }
