+++ src/main/java/com/ggang/be/infra/oauth/OAuthClient.java
@@ -1,6 +1,6 @@
-package com.ggang.be.global.infra.oauth;
+package com.ggang.be.infra.oauth;
 
-import com.ggang.be.global.infra.Auth;
+import com.ggang.be.infra.Auth;
 
 public interface OAuthClient {
     String getAccessToken(String authorizationCode);
